some fixes for using "copy to base dir" together with deep extraction
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@6312 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/optional/jdunrar/UnrarWrapper.java
@@ -316,23 +316,28 @@ public class UnrarWrapper extends Thread implements JDRunnable {
         }
 
     }
-
+    
     private void moveExtractedFilesToBaseDir() {
+        // method fails when used with deep extraction. files that get extracted
+        // in second+ run dont get moved due to missing extractTo path of the
+        // first extraction run. they will stay where they get extracted
         for (ArchivFile f : files) {
-            String[] pathParts = f.getFilepath().split(Regex.escape(File.separator));
-            if (pathParts.length > 1) {
-                File newFile = new File(extractTo.getAbsolutePath() + File.separator + pathParts[0] + File.separator + f.getFile().getName());
-                if (!newFile.exists()) {
-                    f.getFile().renameTo(newFile);
-                    JDLogger.getLogger().warning("Moved file after extraction: " + f + " to " + newFile);
-                    File parentDir = f.getFile().getParentFile();
-                    if (parentDir.isDirectory() && parentDir.listFiles().length == 0) {
-                        parentDir.delete();
-                        JDLogger.getLogger().warning("Deleted empty directory after extraction: " + parentDir);
+            if (JDUnrar.getArchivePartType(f.getFile()) == JDUnrarConstants.NO_RAR_ARCHIVE || JDUnrar.getArchivePartType(f.getFile()) == JDUnrarConstants.NO_START_PART) {
+                String[] pathParts = f.getFilepath().split(Regex.escape(File.separator));
+                if (pathParts.length > 1) {
+                    File newFile = new File(extractTo.getAbsolutePath() + File.separator + pathParts[0] + File.separator + f.getFile().getName());
+                    if (!newFile.exists()) {
+                        f.getFile().renameTo(newFile);
+                        JDLogger.getLogger().warning("Moved file after extraction: " + f + " to " + newFile);
+                        File parentDir = f.getFile().getParentFile();
+                        if (parentDir.isDirectory() && parentDir.listFiles().length == 0) {
+                            parentDir.delete();
+                            JDLogger.getLogger().warning("Deleted empty directory after extraction: " + parentDir);
+                        }
                     }
                 }
             }
-        }
+            }
     }
 
     public Exception getException() {
